[frontend] refacto into useEffect to prevent infinite rerender (#7328)
